Move checkErrors.js to the shared folder
